refactor(mock-data):refactor config file.[MDA-001]
