+++ app/core/pgdb/db.py
@@ -1,6 +1,6 @@
 import databases
 from app.utils.path_builder import build_sql_postgres_connection_string
-from app.config import userNamedb, password, host, port, database_name
+from app.core.config.config import userNamedb, password, host, port, database_name
 
 connectionString = build_sql_postgres_connection_string(
     host=host,
